backend/requirements.txt
Flask==1.0.2
pymongo==3.7.1
python-dotenv==0.10.2
uWSGI==2.0.18